# packages/schemas/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Eyohe OSINT API",
  "description": "Public Intelligence. Connected Evidence. \u2014 local-first OSINT investigation platform.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/cases/{case_id}/entities": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "List Entities",
    "operationId": "list_entities_api_v1_cases__case_id__entities_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Type"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_EntityOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Create Entity",
    "operationId": "create_entity_api_v1_cases__case_id__entities_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EntityCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EntityOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/entities/{entity_id}": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Get Entity",
    "operationId": "get_entity_api_v1_entities__entity_id__get",
    "parameters": [
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Entity Api V1 Entities  Entity Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/cases/{case_id}/relationships": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "List Relationships",
    "operationId": "list_relationships_api_v1_cases__case_id__relationships_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RelationshipOut"
         },
         "title": "Response List Relationships Api V1 Cases  Case Id  Relationships Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Create Relationship",
    "operationId": "create_relationship_api_v1_cases__case_id__relationships_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RelationshipCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RelationshipDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/relationships/{rel_id}": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Get Relationship",
    "operationId": "get_relationship_api_v1_relationships__rel_id__get",
    "parameters": [
     {
      "name": "rel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Rel Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RelationshipDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/relationships/{rel_id}/review": {
   "post": {
    "tags": [
     "entities"
    ],
    "summary": "Review Relationship",
    "operationId": "review_relationship_api_v1_relationships__rel_id__review_post",
    "parameters": [
     {
      "name": "rel_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Rel Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": {
         "type": "string"
        },
        "title": "Payload"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RelationshipDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/cases/{case_id}/graph": {
   "get": {
    "tags": [
     "entities"
    ],
    "summary": "Case Graph",
    "operationId": "case_graph_api_v1_cases__case_id__graph_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     },
     {
      "name": "depth",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 4,
       "minimum": 1,
       "default": 2,
       "title": "Depth"
      }
     },
     {
      "name": "types",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Types"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GraphOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EntityCreate": {
    "properties": {
     "type": {
      "type": "string",
      "title": "Type"
     },
     "value": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 1,
      "title": "Value"
     },
     "label": {
      "type": "string",
      "title": "Label",
      "default": ""
     },
     "attributes": {
      "additionalProperties": true,
      "type": "object",
      "title": "Attributes"
     },
     "evidence_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Evidence Id"
     }
    },
    "type": "object",
    "required": [
     "type",
     "value"
    ],
    "title": "EntityCreate"
   },
   "EntityOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "display_id": {
      "type": "string",
      "title": "Display Id"
     },
     "case_id": {
      "type": "string",
      "format": "uuid",
      "title": "Case Id"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "value": {
      "type": "string",
      "title": "Value"
     },
     "normalized_value": {
      "type": "string",
      "title": "Normalized Value"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "first_seen": {
      "type": "string",
      "format": "date-time",
      "title": "First Seen"
     },
     "last_seen": {
      "type": "string",
      "format": "date-time",
      "title": "Last Seen"
     },
     "confidence": {
      "type": "string",
      "title": "Confidence"
     },
     "source_count": {
      "type": "integer",
      "title": "Source Count"
     },
     "evidence_ids": {
      "items": {},
      "type": "array",
      "title": "Evidence Ids"
     },
     "attributes": {
      "additionalProperties": true,
      "type": "object",
      "title": "Attributes"
     },
     "is_target": {
      "type": "boolean",
      "title": "Is Target"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_id",
     "case_id",
     "type",
     "value",
     "normalized_value",
     "label",
     "first_seen",
     "last_seen",
     "confidence",
     "source_count",
     "evidence_ids",
     "attributes",
     "is_target",
     "is_demo",
     "created_at"
    ],
    "title": "EntityOut"
   },
...
   "GraphEdge": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "display_id": {
      "type": "string",
      "title": "Display Id"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "target": {
      "type": "string",
      "title": "Target"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "confidence": {
      "type": "string",
      "title": "Confidence"
     },
     "review_state": {
      "type": "string",
      "title": "Review State"
     },
     "evidence_count": {
      "type": "integer",
      "title": "Evidence Count"
     },
     "rationale": {
      "type": "string",
      "title": "Rationale"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_id",
     "source",
     "target",
     "type",
     "confidence",
     "review_state",
     "evidence_count",
     "rationale"
    ],
    "title": "GraphEdge"
   },
   "GraphNode": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "display_id": {
      "type": "string",
      "title": "Display Id"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "value": {
      "type": "string",
      "title": "Value"
     },
     "confidence": {
      "type": "string",
      "title": "Confidence"
     },
     "is_target": {
      "type": "boolean",
      "title": "Is Target"
     },
     "evidence_count": {
      "type": "integer",
      "title": "Evidence Count"
     },
     "attributes": {
      "additionalProperties": true,
      "type": "object",
      "title": "Attributes"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_id",
     "type",
     "label",
     "value",
     "confidence",
     "is_target",
     "evidence_count"
    ],
    "title": "GraphNode"
   },
   "GraphOut": {
    "properties": {
     "nodes": {
      "items": {
       "$ref": "#/components/schemas/GraphNode"
      },
      "type": "array",
      "title": "Nodes"
     },
     "edges": {
      "items": {
       "$ref": "#/components/schemas/GraphEdge"
      },
      "type": "array",
      "title": "Edges"
     },
     "stats": {
      "additionalProperties": true,
      "type": "object",
      "title": "Stats"
     }
    },
    "type": "object",
    "required": [
     "nodes",
     "edges"
    ],
    "title": "GraphOut"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Page_EntityOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/EntityOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page",
      "default": 1
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size",
      "default": 50
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[EntityOut]"
   },
...
   "RelationshipCreate": {
    "properties": {
     "source_entity_id": {
      "type": "string",
      "format": "uuid",
      "title": "Source Entity Id"
     },
     "target_entity_id": {
      "type": "string",
      "format": "uuid",
      "title": "Target Entity Id"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "evidence_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "minItems": 1,
      "title": "Evidence Ids"
     },
     "rationale": {
      "type": "string",
      "title": "Rationale",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "source_entity_id",
     "target_entity_id",
     "type",
     "evidence_ids"
    ],
    "title": "RelationshipCreate"
   },
   "RelationshipDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "display_id": {
      "type": "string",
      "title": "Display Id"
     },
     "case_id": {
      "type": "string",
      "format": "uuid",
      "title": "Case Id"
     },
     "source_entity_id": {
      "type": "string",
      "format": "uuid",
      "title": "Source Entity Id"
     },
     "target_entity_id": {
      "type": "string",
      "format": "uuid",
      "title": "Target Entity Id"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "confidence": {
      "type": "string",
      "title": "Confidence"
     },
     "rationale": {
      "type": "string",
      "title": "Rationale"
     },
     "review_state": {
      "type": "string",
      "title": "Review State"
     },
     "review_note": {
      "type": "string",
      "title": "Review Note"
     },
     "attributes": {
      "additionalProperties": true,
      "type": "object",
      "title": "Attributes"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "evidence": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Evidence"
     },
     "source_entity": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EntityOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "target_entity": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/EntityOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "why": {
      "additionalProperties": true,
      "type": "object",
      "title": "Why"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_id",
     "case_id",
     "source_entity_id",
     "target_entity_id",
     "type",
     "confidence",
     "rationale",
     "review_state",
     "review_note",
     "attributes",
     "is_demo",
     "created_at"
    ],
    "title": "RelationshipDetail"
   },
   "RelationshipOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "display_id": {
      "type": "string",
      "title": "Display Id"
     },
     "case_id": {
      "type": "string",
      "format": "uuid",
      "title": "Case Id"
     },
     "source_entity_id": {
      "type": "string",
      "format": "uuid",
      "title": "Source Entity Id"
     },
     "target_entity_id": {
      "type": "string",
      "format": "uuid",
      "title": "Target Entity Id"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "confidence": {
      "type": "string",
      "title": "Confidence"
     },
     "rationale": {
      "type": "string",
      "title": "Rationale"
     },
     "review_state": {
      "type": "string",
      "title": "Review State"
     },
     "review_note": {
      "type": "string",
      "title": "Review Note"
     },
     "attributes": {
      "additionalProperties": true,
      "type": "object",
      "title": "Attributes"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_id",
     "case_id",
     "source_entity_id",
     "target_entity_id",
     "type",
     "confidence",
     "rationale",
     "review_state",
     "review_note",
     "attributes",
     "is_demo",
     "created_at"
    ],
    "title": "RelationshipOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}